+++ project/routers/roles.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 
 from project.database import get_db
@@ -30,6 +30,12 @@ async def get_rol(rol_id: int, db: Session = Depends(get_db)):
 
 @router.post("/", response_model=RolResponseModel, tags=["roles"])
 async def create_rol(rol: RolBaseModel, db: Session = Depends(get_db), usuario_id: int = Depends(get_current_usuario)):
+    usuario = crud.get_usuario(db, usuario_id)
+
+    if usuario.is_admin is False:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Solo usuarios administradores pueden agregar roles.")
+
     new_rol = crud.create_rol(db, rol)
 
     return new_rol
@@ -38,6 +44,12 @@ async def create_rol(rol: RolBaseModel, db: Session = Depends(get_db), usuario_i
 @router.put("/{rol_id}", response_model=RolResponseModel, tags=["roles"])
 async def update_rol(rol: RolBaseModel, rol_id: int, db: Session = Depends(get_db),
                      usuario_id: int = Depends(get_current_usuario)):
+    usuario = crud.get_usuario(db, usuario_id)
+
+    if usuario.is_admin is False:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Solo usuarios administradores pueden modificar roles.")
+
     new_rol = crud.update_rol(db, rol, rol_id)
 
     if new_rol is None:
@@ -48,6 +60,12 @@ async def update_rol(rol: RolBaseModel, rol_id: int, db: Session = Depends(get_d
 
 @router.delete("/{rol_id}", response_model=RolResponseModel, tags=["roles"])
 async def delete_rol(rol_id: int, db: Session = Depends(get_db), usuario_id: int = Depends(get_current_usuario)):
+    usuario = crud.get_usuario(db, usuario_id)
+
+    if usuario.is_admin is False:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Solo usuarios administradores pueden eliminar roles.")
+
     deleted_rol = crud.delete_rol(db, rol_id)
 
     if deleted_rol is None:
